change_roletype_transfer_vtype

+++ supabase/tests/database/schema_roles.sql
@@ -10,7 +10,7 @@ select columns_are('public','roles',
 SELECT has_column( 'roles', 'id' );
 SELECT has_column( 'roles', 'roles' );
 
-select col_type_is('roles','id','bigint');
+select col_type_is('roles','id','smallint');
 select col_type_is('roles','roles','text');
 
 select col_is_pk('roles','id');
